refactor/#307: 공간 메시지 버블 생성 책임 분리
SpaceMessageBubbleFactory 책임
- 버블 랜덤 배치(Entity position setting)
- 텍스트 가공/생성/중앙정렬
- 텍스트에 맞춘 스케일 계산
- 상태 라인 생성/부착
